Added functionality to the gameScene:
firstly the winCheck now works. 
We did a workaround for the view registering when the model changed. we
did by making labels in theModel and then creating labels in the view
using those labels. from there we used a changeListener to register
change. this was our workaround. by changing the label in the model, the
view registered it and performed the action we defined.