Completed the method for randomizing the gameBoard.
It does puzzleSize * 20 moves. for a 3*3 game that'll be 60 moves.
each move is random between the 4 directions the zero can move. it
checks if such a move is possible before passing it to the checkMove()
method. 
This is done by pressing the button. 

timer still needs to implement not going off here